ref(issues): move issues owned endpoints into issues.endpoint folder
