+++ .github/CODEOWNERS
@@ -172,10 +172,6 @@ yarn.lock                                                @getsentry/owners-js-de
 /tests/sentry/incidents/                                  @getsentry/issues
 /tests/acceptance/test_incidents.py                       @getsentry/issues
 
-/src/sentry/api/endpoints/group_events.py                 @getsentry/issues
-/src/sentry/api/endpoints/organization_group_index.py     @getsentry/issues
-/src/sentry/api/endpoints/organization_activity.py        @getsentry/issues
-/src/sentry/api/endpoints/organization_searches.py        @getsentry/issues
 /src/sentry/api/helpers/group_index.py                    @getsentry/issues
 /src/sentry/api/helpers/events.py                         @getsentry/issues
 /src/sentry/api/helpers/group_index/update.py             @getsentry/issues
@@ -327,9 +323,7 @@ static/app/components/events/eventStatisticalDetector/                    @getse
 
 ## Integrations
 /src/sentry/api/endpoints/integrations/                              @getsentry/product-owners-settings-integrations
-/src/sentry/api/endpoints/project_stacktrace_link.py                 @getsentry/issues
 /src/sentry/api/endpoints/organization_integration_repos.py          @getsentry/ecosystem
-/src/sentry/api/endpoints/organization_release_previous_commits.py   @getsentry/issues
 /src/sentry/api/endpoints/sentry_app/                                @getsentry/product-owners-settings-integrations
 /src/sentry/api/endpoints/project_rule*.py                           @getsentry/alerts-notifications
 /src/sentry/api/serializers/models/rule.py                           @getsentry/alerts-notifications
@@ -512,7 +506,6 @@ static/app/components/events/eventStatisticalDetector/                    @getse
 /src/sentry/search/snuba/                           @getsentry/issues
 /static/app/views/issueList                                  @getsentry/issues
 /static/app/views/organizationGroupDetails                   @getsentry/issues
-/src/sentry/api/endpoints/source_map_debug.py                @getsentry/issues
 /src/sentry/api/helpers/source_map_helper.py                  @getsentry/issues
 /src/sentry/api/endpoints/actionable_items.py                @getsentry/issues
 /src/sentry/api/helpers/actionable_items_helper.py                  @getsentry/issues
