+++ src/sentry/integrations/utils/stacktrace_link.py
@@ -14,7 +14,7 @@ from sentry.shared_integrations.exceptions import ApiError
 from sentry.utils.event_frames import EventFrame
 
 if TYPE_CHECKING:
-    from sentry.api.endpoints.project_stacktrace_link import StacktraceLinkContext
+    from sentry.issues.endpoints.project_stacktrace_link import StacktraceLinkContext
 
 logger = logging.getLogger(__name__)
