+++ tests/sentry/ratelimits/utils/test_get_ratelimit_key.py
@@ -3,8 +3,8 @@ from django.test import RequestFactory
 from rest_framework.permissions import AllowAny
 
 from sentry.api.base import Endpoint
-from sentry.api.endpoints.organization_group_index import OrganizationGroupIndexEndpoint
 from sentry.auth.system import SystemToken
+from sentry.issues.endpoints.organization_group_index import OrganizationGroupIndexEndpoint
 from sentry.models.apitoken import ApiToken
 from sentry.models.user import User
 from sentry.ratelimits import get_rate_limit_config, get_rate_limit_key
